third_party/WebKit/LayoutTests/shadow-dom/crashes/add-host-child-element-crash.html

<!DOCTYPE html>
<script src='../../resources/testharness.js'></script>
<script src='../../resources/testharnessreport.js'></script>
<div id="host"></div>
<script>
test(() => {
  const host = document.querySelector('#host');
  const shadowRoot = host.attachShadow({'mode': 'open'});
  const slot = document.createElement('slot');
  shadowRoot.appendChild(slot);
  host.appendChild(document.createElement('div'));
  document.body.offsetLeft;
  host.appendChild(document.createElement('div'));
  document.body.offsetLeft;
  assert_true(true, 'This test must run without crash.');
});
</script>
